+++ src/WINNT/afsusrmgr/window.h
@@ -18,7 +18,7 @@
 
 BOOL CALLBACK Main_DialogProc (HWND hDlg, UINT msg, WPARAM wp, LPARAM lp);
 
-void Main_PrepareTabChild (int iTabNew = -1);
+void Main_PrepareTabChild (size_t iTabNew = -1);
 
 void Main_SetMenus (void);
 void Main_SetViewMenus (HMENU hm);
